requirements.txt
asgiref==3.4.1
Django==3.2.7
Pillow==8.3.2
python-json-logger==2.0.2
pytz==2021.1
sqlparse==0.4.2